Add benchmark methods for JustMockLite
added JustMock, but currently running into issues during testing (some tests take a long time to finish)

added try/finally blocks around JustMock code and put Mock.Reset() in the finally blocks - seems to help.

pushed Reset() to the beginning of each JustMock method, that seems to be the trick. Also updated the benchmark methods to "JustMockLite" and changed the alias to "JustMock".